Music Additions

+++ src/main/java/com/beanbeanjuice/command/music/RepeatCommand.java
@@ -5,6 +5,7 @@ import com.beanbeanjuice.utility.command.CommandContext;
 import com.beanbeanjuice.utility.command.ICommand;
 import com.beanbeanjuice.utility.command.usage.Usage;
 import com.beanbeanjuice.utility.command.usage.categories.CategoryType;
+import com.beanbeanjuice.utility.command.usage.types.CommandType;
 import com.beanbeanjuice.utility.lavaplayer.GuildMusicManager;
 import com.beanbeanjuice.utility.lavaplayer.PlayerManager;
 import net.dv8tion.jda.api.EmbedBuilder;
@@ -27,6 +28,16 @@ public class RepeatCommand implements ICommand {
     @Override
     public void handle(CommandContext ctx, ArrayList<String> args, User user, GuildMessageReceivedEvent event) {
 
+        String commandName = args.get(0).toLowerCase();
+
+        if (!commandName.equals("song") && !commandName.equals("playlist")) {
+            event.getChannel().sendMessage(BeanBot.getGeneralHelper().errorEmbed(
+                    "Incorrect Command Term",
+                    "You must use the command term `song` or `playlist` not `" + commandName + "`."
+            )).queue();
+            return;
+        }
+
         event.getMessage().delete().queue();
         BeanBot.getGuildHandler().getCustomGuild(event.getGuild()).setLastMusicChannel(event.getChannel());
 
@@ -48,13 +59,36 @@ public class RepeatCommand implements ICommand {
             }
 
             GuildMusicManager musicManager = PlayerManager.getInstance().getMusicManager(ctx.getGuild());
-            final boolean newRepeating = !musicManager.scheduler.repeating;
-            musicManager.scheduler.repeating = newRepeating;
 
-            if (newRepeating) {
-                event.getChannel().sendMessage(successEmbed("Song repeating has now been turned on.")).queue();
-            } else {
-                event.getChannel().sendMessage(successEmbed("Song repeating has now been turned off.")).queue();
+            if (commandName.equals("song")) {
+                final boolean newRepeating = !musicManager.scheduler.songRepeating;
+                musicManager.scheduler.songRepeating = newRepeating;
+
+                if (newRepeating) {
+                    event.getChannel().sendMessage(successEmbed("Song repeating has now been turned on.")).queue();
+                } else {
+                    event.getChannel().sendMessage(successEmbed("Song repeating has now been turned off.")).queue();
+                }
+            }
+
+            if (commandName.equals("playlist")) {
+
+                if (!musicManager.scheduler.playlistRepeating() && musicManager.scheduler.queue.isEmpty()) {
+                    event.getChannel().sendMessage(BeanBot.getGeneralHelper().errorEmbed(
+                            "Cannot Repeat Playlist",
+                            "Cannot repeat the playlist as the playlist is currently empty."
+                    )).queue();
+                    return;
+                }
+
+                final boolean newRepeating = !musicManager.scheduler.playlistRepeating();
+                musicManager.scheduler.setPlaylistRepeating(newRepeating);
+
+                if (newRepeating) {
+                    event.getChannel().sendMessage(successEmbed("Playlist repeating has now been turned on.")).queue();
+                } else {
+                    event.getChannel().sendMessage(successEmbed("Playlist repeating has now been turned off.")).queue();
+                }
             }
 
 
@@ -100,7 +134,9 @@ public class RepeatCommand implements ICommand {
 
     @Override
     public Usage getUsage() {
-        return new Usage();
+        Usage usage = new Usage();
+        usage.addUsage(CommandType.TEXT, "song/playlist", true);
+        return usage;
     }
 
     @Override
